# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PharmaCare Digital Pharmacy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patient/consultations:
    get:
      operationId: getPatientConsultations
      tags: [patients]
      summary: Get all consultations for the authenticated patient
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Patient consultations
          content:
            application/json:
              schema:
                type: object
                properties:
                  consultations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Consultation"
                required:
                  - consultations
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    Consultation:
      type: object
      properties:
        id:
          type: string
        patientName:
          type: string
        patientEmail:
          type: string
        patientAge:
          type: integer
        patientSex:
          type: string
          enum: [male, female, other]
        conditionId:
          type: string
        conditionName:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, more_info_needed, referred, red_flag, patient_responded, cancelled]
        answers:
          type: object
          additionalProperties: true
        hasRedFlag:
          type: boolean
        hasPhoto:
          type: boolean
        photoUrls:
          type: array
          items:
            type: string
        pharmacistNote:
          type: string
          nullable: true
        prescription:
          type: string
          nullable: true
        prescriptionItems:
          type: array
          items:
            $ref: "#/components/schemas/PrescriptionItem"
          nullable: true
        referralInfo:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        # ── Extended fields (GPhC compliance + intake data) ──
        gpName:
          type: string
          nullable: true
        gpSurgery:
          type: string
          nullable: true
        gpAddress:
          type: string
          nullable: true
        gpPhone:
          type: string
          nullable: true
        hasRegularGp:
          type: boolean
          nullable: true
        consentShareWithGp:
          type: boolean
          nullable: true
        consentToTreatment:
          type: boolean
          nullable: true
        consentToDelivery:
          type: boolean
          nullable: true
        consentDataProcessing:
          type: boolean
          nullable: true
        identityVerificationMethod:
          type: string
          nullable: true
        identityVerificationRef:
          type: string
          nullable: true
        deliveryAddress:
          type: string
          nullable: true
        deliveryAddressLine1:
          type: string
          nullable: true
        deliveryAddressLine2:
          type: string
          nullable: true
        deliveryCity:
          type: string
          nullable: true
        deliveryPostcode:
          type: string
          nullable: true
        preferredDeliveryMethod:
          type: string
          nullable: true
        allergies:
          type: string
          nullable: true
        currentMedications:
          type: string
          nullable: true
        medicalHistory:
          type: string
          nullable: true
        isPregnant:
          type: boolean
          nullable: true
        bmi:
          type: number
          nullable: true
        verifiedHeightCm:
          type: number
          nullable: true
        verifiedWeightKg:
          type: number
          nullable: true
        heightCm:
          type: number
          nullable: true
        weightKg:
          type: number
          nullable: true
        riskFlags:
          type: array
          items:
            type: string
          nullable: true
        riskCategory:
          type: string
          nullable: true
        clinicalDecisionRationale:
          type: string
          nullable: true
        reviewedBy:
          type: string
          nullable: true
        deliveryCarrier:
          type: string
          nullable: true
        deliveryTrackingNumber:
          type: string
          nullable: true
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        dispatchedAt:
          type: string
          format: date-time
          nullable: true
        previousConsultationId:
          type: string
          nullable: true
          description: ID of the previous consultation this one is a repeat / follow-up of.
      required:
        - id
        - patientName
        - patientEmail
        - patientAge
        - patientSex
        - conditionId
        - conditionName
        - status
        - answers
        - hasRedFlag
        - hasPhoto
        - createdAt
        - updatedAt
...
    PrescriptionItem:
      type: object
      properties:
        name:
          type: string
        strength:
          type: string
        form:
          type: string
        quantity:
          type: string
        sig:
          type: string
        duration:
          type: string
        notes:
          type: string
          nullable: true
      required: [name, strength, form, quantity, sig, duration]
...